refactor: 검색 로직 패키지 분리
